fix: resolve CodeRabbit review findings across CI, runtime safety, and docs
Critical CI fixes:
- Fix LICENSE copyright from "Stringy Contributors" to correct project
- Fix README license reference from MIT to Apache 2.0 (matches LICENSE)
- Fix Codecov slug pointing to wrong repository (StringyMcStringFace)
- Fix coverage upload path (target/lcov.info → lcov.info)
- Fix release.yml SBOM upload typo (output → outputs)
- Remove non-functional docs.rs badge from README

Runtime safety fixes:
- Prevent par_chunks(0) panic in streaming.rs for empty configs
- Fix u16 overflow in vlan.rs total_vlans calculation (cast before math)
- Add empty-vec guard in vpn.rs DNS selection to prevent range panic

Code/config quality:
- Fix deny.toml project name (DaemonEye → OPNsense Config Faker)
- Fix deny.toml exclude-dev/include-dev conflict
- Fix AGENTS.md INSTA_UPDATE=auto → always, Rust edition 2021 → 2024
- Add null checks in mermaid-init.js getElementById calls
- Fix justfile recipe name typo (lint:actions → lint-actions)
- Fix PR template trailing backslash consistency

Documentation accuracy:
- Fix VLAN range 1-4094 → 10-4094 across 4 docs files
- Fix Rust version 1.70+ → 1.85+ in installation docs
- Fix broken relative links in 4 docs files
- Remove fake RUST_GC_THRESHOLD, sudo cargo run, invalid --registry flag
- Remove non-existent CLI flags from output-formats docs
- Fix JS reserved word 'interface' in examples
- Fix cargo-llvm-cov listed as dev-dependency
- Fix broken nested code fences in migration docs

+++ src/generator/vlan.rs
@@ -675,7 +675,7 @@ pub fn generate_vlan_configurations_from_ranges(
     // Calculate total number of VLANs for progress tracking and pre-allocation
     let total_vlans: u32 = vlan_ranges
         .iter()
-        .map(|(start, end)| (*end - *start + 1) as u32)
+        .map(|(start, end)| u32::from(*end) - u32::from(*start) + 1)
         .sum();
     let mut configs = Vec::with_capacity(total_vlans as usize);
     let mut processed = 0u64;
@@ -716,7 +716,7 @@ pub fn generate_vlan_configurations_from_ranges_with_wan(
     // Calculate total number of VLANs for progress tracking and pre-allocation
     let total_vlans: u32 = vlan_ranges
         .iter()
-        .map(|(start, end)| (*end - *start + 1) as u32)
+        .map(|(start, end)| u32::from(*end) - u32::from(*start) + 1)
         .sum();
     let mut configs = Vec::with_capacity(total_vlans as usize);
     let mut processed = 0u64;
